amem: fix integer overflow above 262143Hz
The variable 'rate' in aout_sys_t had 18bits, which was
not enough to represent values greater than 262143.

This patch increases the number of bits reserved for
the sample rate while keeping enough bits for 'channels'
to be able to represent up to 1023;

This patch also fixes an inconsistency in the number
of supported channels. According to source code, where
channel mapping is configured, amem only supports 8
channels.

Signed-off-by: Farid Hammane <farid.hammane@gmail.com>